+++ spec/rails_app/config/initializers/devise-multi_email.rb
@@ -1,5 +1,6 @@
 
 Devise::MultiEmail.configure do |config|
-  # specify this for backwards-compatibility
-  config.primary_email_method_name = :primary_email_record
+  #config.configure_autosave = false
+  #config.primary_email_method_name = :primary_email
+  #config.only_login_with_primary_email = false
 end